Made italics light, made HTML more readable
Finally figured out how to make the italics light. I had to import
a font (I went with the Roboto that they used for the example) and after
copying the link into the html header, the font-weights worked. I also
swapped out some of the divs I used for text into paragraph tags. I need
to be better about not using divs for so many things in html, as it is
not best practice according to all of our the extracurricular videos
I've been watching.